Use quoted includes in headers.
Compilers like Clang issue a compile error when using angle brackets for
these include files. This due to the location of the header files.
On Arch Linux, the include is done using <PCSC/winscard.h>.

When now using angle brackets in winscard.h, the compiler can't find
pcsclite.h because it is not in the default include path, it is in
<PCSC/pcsclite.h>, not <pcsclite.h>.

When using quotes, the compiler searches in the directory of the
including file first, where it can find the header.