SCardGetStatusChange: speedup the case DISABLE_AUTO_POWER_ON
If DISABLE_AUTO_POWER_ON is defined then we do not waste time waiting
for the card to be powered. The card may not be powered before long.

Since we wait PCSCLITE_STATUS_POLL_RATE (400000 µs) for each card, the
delay can be very long for an 8 slots reader (sysmoOCTSIM):
400ms * 8 = 3.2 seconds

or very very long for a 96 slots reader (sysmoSIMBANK-96):
400ms * 96 = 38.4 seconds