Makefile.am: do not include pcsclite.h in archive
The file src/PCSC/pcsclite.h is generated from src/PCSC/pcsclite.h.in by
./configure
No need to include it in the tarball archive.

Also no need to patch the .h in additon to the .h.in to update the value
of PCSCLITE_MAX_READERS_CONTEXTS as done in the Fedora package.