libudev: do not fail removal if the device has no parent
With some USB devices (like Yubikey) a new USB bus is created when the
device is inserted. The USB bus is then destroyed when the token is
removed.
The USB device has no parent any more and the call to
udev_device_get_parent_with_subsystem_devtype() fails. The device is
then NOT removed and you get a loop of:

 ccid_usb.c:886:WriteUSB() write failed (3/2): LIBUSB_ERROR_NO_DEVICE
 src/ifdwrapper.c:364:IFDStatusICC() Card not transacted: 617
 src/eventhandler.c:336:EHStatusHandlerThread() Error communicating to: Yubico YubiKey OTP+FIDO+CCID 00 00

Closes: https://github.com/LudovicRousseau/PCSC/issues/125
Thanks to oddlama for the bug report