Fix SCardReleaseContext failure on orphan handles
Fix returning a spurious error from SCardReleaseContext() for a
correct SCARDCONTEXT, in case there was an orphan SCARDHANDLE left after
disconnecting a reader.

The error was caused by the fact that SCardReleaseContext attempts to do
SCardDisconnect for all active SCARDHANDLEs stored in SCARDCONTEXT,
which fails when the handle was already deleted by the hotplug code from
READER_CONTEXT or the reader got zero reference count.