Add '--disable-polkit' option
On systems with an encrypted root drive, systemd-cryptsetup may be used
to handle decrypting the disk with a PKCS#11 token via pcscd. For this
to work however, pcscd has to run in the initramfs to handle
communications with the smartcard.

To avoid having to also add polkitd to the initramfs, or alternatively
to avoid having 2 separate versions of pcscd installed on the same
system, add a commandline flag to disable polkit authentication for
these cases.

https://salsa.debian.org/rousseau/PCSC/-/merge_requests/4