pcscd: allow configuring the auto-exit timeout
The auto-exit feature is hard-coded to exit after 60 seconds. But users
who bump the power-down timeout to something higher will want pcscd to
keep running, so that the cards may continue to consider it a single
session (and so not require the PIN to be input again).

Let's add an optional timeout argument to let them set it. We could
perhaps guess at this situation by pushing the auto-exit timeout if the
a power-down timeout is still running. But it is simplest to just let
the user configure the two independently.

Signed-off-by: Jeff King <peff@peff.net>