pcscd: allow configuring the power-down timeout
If auto power-down is enabled, the grace period timeout of 5 seconds is
hard-coded. There is no way to disable the feature nor change the
timeout without recompiling.

However, users may want a longer timeout in order to keep their
"session" with the card open (even when no client is connected).  For
example, the yubikey plugin for the "age" tool has no long-running
process; it connects and disconnects each time "age" is run. But using a
yubikey PIV which is configurd to require the PIN only once per session
is useless, as the session times out after only 5 seconds (after which
the user must provide the PIN again).

This patch provides a command-line option to set the power-down timeout
in seconds. If no option is given, we'll continue to use 5 seconds as
the default.

Signed-off-by: Jeff King <peff@peff.net>