libpcscspy: fix a crash with NULL pointers
If the application uses pioRecvPci parameter set to NULL in
SCardTransmit() then the spy library crashed.
But NULL is a valid value for this parameter.

We also handle tha case pioSendPci set to NULL even if this case will
return an error, but the spy library also crashed.

Thanks to Amreo for the bug report
Fixes: https://github.com/LudovicRousseau/PCSC/issues/159