SCardConnect() & SCardReconnect(): restrict the protocol used
We should not try to use a protocol that is not in the preferred ones
from the application.
If the application requests to use T=1 only but the card does support T=0
& T=1, and the reader does support T=0 only we should NOT try to use T=0
but fail instead.

SCardConnect() & SCardReconnect() will now return SCARD_E_PROTO_MISMATCH.

This should match the Windows behavior.

Thanks to Dmitry for the bug report
" Erroneously success of SCardConnect in case of T=0 reader #162 "
https://github.com/LudovicRousseau/PCSC/issues/162