PHSetProtocol(): do not use a protocol not selected
If the application requests T=1 for a card that support both T=0 and T=1
and that the PTS using T=1 fails then we should not try to use T=0.

This is the case if the reader does not support T=1 but only T=0.

Thanks to Dmitry for the bug report
" Erroneously success of SCardConnect in case of T=0 reader #162 "
https://github.com/LudovicRousseau/PCSC/issues/162