Use a fixed date in pcsc-spy.1 manpage
The manpage pcsc-spy.1 is generated by pod2an from pcsc-spy.pod. The
date included in the manpage is the current date of generation.
This caused problems when installing libpcsclite-dev for 2 different
architectures at the same time.

Thanks to Francois Gouget for the bug report
"Bug#1064636: libpcsclite-dev: The i386 libpcsclite-dev 2.0.1-1+b1 package conflicts with the amd64 one"
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1064636