Stop depending on libsystemd for socket activation
As libsystemd is usually only available on systems using systemd, it is
not possible to use socket activation with pcscd on systems which do not
use systemd.

The socket activation mechanism used by systemd is relatively simple,
and re-implementing it is trivial. Socket activation can be convenient
even when systemd is not used.

This patch removes the optional libsystemd dependency and re-implements
the core functionality which was originally depended on in order to
enable socket-activation on all systems assuming the circumstances are
correct.

This change maintains full compatibility with systemd.