hotplug_libusb: fix cleanup for IFD_GENERATE_HOTPLUG
If IFD_GENERATE_HOTPLUG is defined by all USB drivers:
    <key>ifdCapabilities</key>
    <string>0x00000001</string>

    <!-- Possible values for ifdCapabilities bits
    1: IFD_GENERATE_HOTPLUG
       plugging the reader calls pcscd \-\-hotplug

    Default value is 0x00000000
    -->

then the hotplug mechanism does not use polling to rescan the USB bus
but waits for pcscd to be called with "--hotplug" to rescan the USB bus
and serial configurations.

This case was not handled correctly on pcscd exit and the
HPEstablishUSBNotifications thread was not terminated.