SCardGetStatusChange(): add the number of reader events
The special reader "\\?PnP?\Notification" now returns the number of
reader events in the 16 high bits of the .dwEventState field.

The goal is to be able to detect a reader change between two calls to
SCardGetStatusChange(). If a new reader has been connected or a reader
has been disconnected then SCardGetStatusChange() will return immediatly
with SCARD_STATE_CHANGED set for "\\?PnP?\Notification".