Support udev PCSCLITE_IGNORE property to filter readers
Adds support for filtering readers by checking whether the udev property
PCSCLITE_IGNORE is set to "1" for the device being probed.

This is similar to the existing filtering support provided by
PCSCLITE_FILTER_IGNORE_READER_NAMES, but allows using the full range of udev
rules to match specific devices. For example, if you have several Yubikeys and
one is attached to a specific USB port that is being forwarded to a virtual
machine (viz. it should not be touched by programs on the host machine), this
facility allows setting a rule to keep pcscd away from that Yubikey. Yubikeys
attached to other ports keep working normally.

Example rule that targets a Yubikey attached to a specific USB port and causes
pcscd to ignore that device:

ACTION!="remove|unbind", SUBSYSTEM=="usb", ATTR{idVendor}=="1050",
ATTR{idProduct}=="0406", KERNEL=="1-2.3", ENV{PCSCLITE_IGNORE}="1"

The KERNEL match targets a USB device attached to port 3 of a hub on port 2 of
USB bus 1, and rule sets property PCSCLITE_IGNORE to cause the device to be
ignored.

Signed-off-by: Valtteri Vuorikoski <vuori@notcom.org>