meson: respect builtin meson default_library option
pcsclite can be built both statically or shared. It has to be built
differently to support static usage due to the delegate mechanism. In
order to handle this, separate calls are made to static_library() and
shared_library(), which sidestep meson's usual handling for selecting
the correct requested build variant.

Solve this by looking up the builtin option and choosing whether or not
to build each library. This means people intending to only ship the
shared library, no longer spend time compiling the static one -- nor
have to manually delete it after it is installed by `meson install`.

Fixes a stray static library appearing on Gentoo (which builds with
default_library=shared).

Ref: https://github.com/LudovicRousseau/PCSC/issues/216
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>