Add backward version support on the server side
The server now also supports clients using the previous communication
protocol 4:4 in addition to 4:5.

Protocol 4:5 was introduced with pcsc-lite 2.3.0 in Aug 2024.
Protocol 4:4 was introduced with pcsc-lite 1.8.24 in Oct 2018.

Protocol 4:5 adds the command CMD_GET_READER_EVENTS. This command will
not be used by older clients. They will just work as if the server was
also using protocol version 4:4.